Ресурсы пользователей и проектов защищены аутентификацией

+++ api/resources/ProjectResource.py
@@ -1,6 +1,7 @@
-from flask import jsonify
+from flask import jsonify, g
 from flask_restful import abort, Resource
 
+from api.auth import token_auth
 from api.data import db_session
 from api.resources.parsers import project_parser_for_adding, project_parser_for_updating
 from api.data.project import Project
@@ -17,29 +18,48 @@ def abort_if_project_not_found(func):
     return new_func
 
 
+def check_if_user_is_a_member(func):
+    def new_func(self, project_id):
+        if project_id not in map(lambda x: x.id, g.current_user.projects):
+            abort(403)
+        return func(self, project_id)
+
+    return new_func
+
+
 class ProjectResource(Resource):
     @abort_if_project_not_found
+    @token_auth.login_required
+    @check_if_user_is_a_member
     def get(self, project_id):
         session = db_session.create_session()
         project = session.query(Project).get(project_id)
+        if project not in g.current_user.projects:
+            abort(403)
         return jsonify({
             'project': project.to_dict(only=('team_leader_id', 'project_name', 'title', 'description', 'reg_date')),
             'users': [item.id for item in project.users]})
 
     @abort_if_project_not_found
+    @token_auth.login_required
     def delete(self, project_id):
         session = db_session.create_session()
         project = session.query(Project).get(project_id)
+        if project.creator != g.current_user:
+            abort(403)
         session.delete(project)
         session.commit()
         return jsonify({'success': True})
 
     @abort_if_project_not_found
+    @token_auth.login_required
     def put(self, project_id):
         args = project_parser_for_updating.parse_args(strict=True)  # Вызовет ошибку, если запрос 
         # будет содержать поля, которых нет в парсере
         session = db_session.create_session()
         project = session.query(Project).get(project_id)
+        if project.creator != g.current_user:
+            abort(403)
         for key, value in args.items():
             if value is not None:
                 exec(f"project.{key} = '{value}'")
@@ -48,22 +68,22 @@ class ProjectResource(Resource):
 
 
 class ProjectListResource(Resource):
+    @token_auth.login_required
     def get(self):
-        session = db_session.create_session()
-        projects = session.query(Project).all()
         return jsonify({
             'projects': [
                 {
                     'project': project.to_dict(
                         only=('team_leader_id', 'project_name', 'title', 'description', 'reg_date')),
-                    'users': [item.id for item in project.users]} for project in projects],
+                    'users': [item.id for item in project.users]} for project in g.current_user.projects],
         })
 
+    @token_auth.login_required
     def post(self):
         args = project_parser_for_adding.parse_args(strict=True)
         session = db_session.create_session()
         project = Project(
-            team_leader_id=args['team_leader_id'],
+            team_leader_id=g.current_user.id,
             project_name=args['project_name'],
             title=args['title'],
             description=args['description']
